service/requirements.txt
Django==3.1.7
Pillow==8.4.0
boto3==1.19.0
cryptography==2.8
django-cors-headers==3.7.0
django-storages==1.12.2
djangorestframework==3.12.4
graphene-django==2.15.0
gunicorn==20.1.0